+++ Plugins/iOS/PushNotificationManagerWrapper.mm
@@ -7,5 +7,6 @@ extern "C" {
     
     void RegisterUnityCallbackForNotificationReceived(NOTIFICATION_CALLBACK callback) {
         [[PushNotificationManager sharedInstance] setNotificationCallback:callback];
+        [[PushNotificationManager sharedInstance] flushNotificationBuffer];
     }
 }