+++ Runtime/PushNotifications.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using UnityEngine;
 
 [assembly: InternalsVisibleTo("Unity.Services.PushNotifications.Tests")]
+[assembly: InternalsVisibleTo("Unity.Services.PushNotifications.Editor")]
 
 namespace Unity.Services.PushNotifications
 {
@@ -58,7 +59,28 @@ namespace Unity.Services.PushNotifications
         /// </summary>
         /// <param name="settings">A PushNotificationSettings object with the settings for the SDK. See the documentation on that class for more information.</param>
         /// <returns>(Asynchronously via a Task) The device token as a string.</returns>
+        public static Task<string> RegisterForPushNotificationsAsync()
+        {
+            PushNotificationSettings settings = PushNotificationSettings.GetAssetInstance();
+            return RegisterForPushNotificationsInternal(settings);
+        }
+
+        /// <summary>
+        /// (Deprecated - use the parameterless version and set settings in the editor GUI instead. Using this method will overwrite any settings you have in the GUI settings interface).
+        /// Registers for push notifications with the appropriate mechanism for the current platform.
+        ///
+        /// This method will automatically handle platform specific intricacies of getting a push notification token, and will
+        /// send the appropriate analytics events to Unity Analytics 2.
+        /// </summary>
+        /// <param name="settings">A PushNotificationSettings object with the settings for the SDK. See the documentation on that class for more information.</param>
+        /// <returns>(Asynchronously via a Task) The device token as a string.</returns>
+        [Obsolete("Settings should now be configured in the Editor's Project Settings, and then register using the parameterless version of RegisterForPushNotificationsAsync", false)]
         public static Task<string> RegisterForPushNotificationsAsync(PushNotificationSettings settings)
+        {
+            return RegisterForPushNotificationsInternal(settings);
+        }
+
+        internal static Task<string> RegisterForPushNotificationsInternal(PushNotificationSettings settings) 
         {
             if (!isInitialised)
             {
@@ -67,7 +89,7 @@ namespace Unity.Services.PushNotifications
 #if UNITY_IOS
             return s_IOSPushNotifications.RegisterForPushNotificationsAsync();
 #elif UNITY_ANDROID
-            return s_AndroidPushNotifications.RegisterForPushNotificationsAsync(settings.AndroidApiKey, settings.AndroidSenderId, settings.AndroidApplicationId, settings.AndroidProjectId);
+            return s_AndroidPushNotifications.RegisterForPushNotificationsAsync(settings.androidApiKey, settings.androidSenderId, settings.androidApplicationId, settings.androidProjectId);
 #else
             Debug.Log("Push notifications are not supported on this platform at this time, returning an empty push token");
             return Task.FromResult("");